% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{add.fx.to.matrix}
\alias{add.fx.to.matrix}
\title{Add a true effect to a matrix of effect sizes (or generate a new one).}
\usage{
add.fx.to.matrix(
  samples,
  species,
  which.species,
  which.samples,
  mtx = 0,
  fx = 1
)
}
\arguments{
\item{samples}{Vector of strings representing the sample names.}

\item{species}{Vector of strings representing species names.}

\item{which.species}{Vector of strings giving subset of species to be affected.}

\item{which.samples}{Vector of strings giving which samples to be affected.}

\item{mtx}{Optionally, an existing matrix (if not provided, generate a new
one).}

\item{fx}{Size of effect to be added.}
}
\value{
A new effect size matrix.
}
\description{
Add a true effect to a matrix of effect sizes (or generate a new one).
}
\keyword{internal}
