% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{add.sig.descs}
\alias{add.sig.descs}
\title{Add gene descriptions to significant results; return in a tibble.}
\usage{
add.sig.descs(phy.with.sigs, pos.sig, gene.to.fxn)
}
\arguments{
\item{phy.with.sigs}{Character vector giving the phyla with significant hits.}

\item{pos.sig}{List of character vectors, one per phylum, of significant hits.}

\item{gene.to.fxn}{Data frame used to annotate genes to functions.}
}
\value{
A single data frame of all significant results plus descriptions.
}
\description{
Add gene descriptions to significant results; return in a tibble.
}
