% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{alt.multi.enrich}
\alias{alt.multi.enrich}
\title{An alternative way to get the enrichment table, using a tbl of results
instead of separate inputs for significant results, effect signs, etc.}
\usage{
alt.multi.enrich(results, mappings, dirxn = 1, qcuts = c(strong = 0.05,
  med = 0.1, weak = 0.25), future = FALSE)
}
\arguments{
\item{results}{A tidyverse tbl of results with at least the following
columns: "phylum", "gene", "effect.size", and "q.value" (if only
"p.value" is present, first apply function \code{tbl.result.qvs}).}

\item{mappings}{List of data.frames giving gene-to-gene-set mappings.}

\item{dirxn}{Count only genes with this effect sign as significant.}
}
\value{
A tbl giving Fisher's test p-values, q-values, effect sizes, and overlaps.
}
\description{
An alternative way to get the enrichment table, using a tbl of results
instead of separate inputs for significant results, effect signs, etc.
}
