% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{annotate.nested}
\alias{annotate.nested}
\title{Transform nested lists into a data frame.}
\usage{
annotate.nested(nested, summarize = NULL, n = NULL, n.names = NULL,
  stop.at = 0)
}
\arguments{
\item{nested}{A nested list.}

\item{summarize}{If NULL, indicates that no summarizing should be done on the
most basic list elements; if a function or closure, this function or
closure will be applied to the most basic list elements (for example,
\code{mean} would take the mean of vectors at the "bottom" of the nested
list), and the return value are incorporated into the final table instead
of the originals.}

\item{n}{An optional vector of values to bind to the result table.}

\item{n.names}{An optional vector of names for the columns in the final table.}

\item{stop.at}{Stop at this depth and summarize.}
}
\value{
A 2D data frame representation of the (summarized) values in the
    nested lists.
}
\description{
\code{annotate.nested} takes a nested list (i.e., a list of lists, each of
which lists may also be a list of lists), and transforms it into a tabular
structure. This is useful for visualizing and processing the results of
nested \code{lapply} statements, such as those that might be used when
performing a grid search across multiple parameters.
}
