% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{augment_with_enrichments}
\alias{augment_with_enrichments}
\title{Add enrichments to a core phylogenize object}
\usage{
augment_with_enrichments(core)
}
\arguments{
\item{core}{The named list obtained from running \code{phylogenize_core()}.}
}
\value{
The input core object with an \code{enr_tbls} element added.
}
\description{
Add enrichment tables after the fact to a phylogenize core object.
}
