% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{capwords}
\alias{capwords}
\title{Capitalize the first letter of a word/vector of words.}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{Word or vector of words.}

\item{strict}{If TRUE, also force non-initial letters to be lowercase.}
}
\description{
Capitalize the first letter of a word/vector of words.
}
