% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{check.process.metadata}
\alias{check.process.metadata}
\title{Check and process metadata}
\usage{
check.process.metadata(metadata, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{metadata}{A data frame of metadata with environment, dataset, and
sample columns corresponding to the resolved options.}
}
\value{
A data frame of metadata, with environment and
    dataset columns converted to factors, *unless* calculating correlation
    in which case environment column will be cast as numeric. The environment
    factor will be automatically re-leveled so that \code{envir} is last, so
    that it is guaranteed to appear in the output of differential abundance 
    estimators.
}
\description{
\code{check.process.metadata} is used to make sure that the metadata
satisfies the requirements specified by the resolved options and to make sure
that the metadata are of the correct type.
}
\details{
Some particularly relevant options are:
\describe{
  \item{env_column}{Name of metadata column containing environment
    annotations.}
  \item{envir}{Environment of interest.}
  \item{dset_column}{Name of metadata column containing dataset annotations.}
  \item{single_dset}{Boolean. If true, will assume that all samples come from
  a single dataset called \code{dset1} no matter what, if anything, is in
  \code{dset_column}.}
}
}
