% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{clean.pheno}
\alias{clean.pheno}
\title{Remove taxa from a phenotype that aren't in our trees and gene matrices
(usually only necessary for testing).}
\usage{
clean.pheno(phenotype, pz.db)
}
\arguments{
\item{phenotype}{A quantitative phenotype (named numeric vector).}

\item{pz.db}{A database.}
}
\value{
A phenotype with only the measurements represented in the database.
}
\description{
Remove taxa from a phenotype that aren't in our trees and gene matrices
(usually only necessary for testing).
}
