% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{cluster.load.pkg}
\alias{cluster.load.pkg}
\title{Load the \code{phylogenize} package, either using \code{library} or
\code{devtools} as appropriate.}
\usage{
cluster.load.pkg(cl, devel, pkgdir = "package/phylogenize")
}
\arguments{
\item{cl}{A \code{parallel} cluster}

\item{devel}{Boolean; if TRUE, use \code{devtools}, otherwise use
\code{library}}

\item{pkgdir}{Optional string giving path to package; used with
\code{devtools} only}
}
\description{
Load the \code{phylogenize} package, either using \code{library} or
\code{devtools} as appropriate.
}
\keyword{internal}
