% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{colnorm}
\alias{colnorm}
\title{Normalize a matrix by columns (i.e., force them to sum to one).}
\usage{
colnorm(mtx)
}
\arguments{
\item{mtx}{Numeric matrix.}
}
\value{
A numeric matrix where each column has been divided by its sum.
}
\description{
Normalize a matrix by columns (i.e., force them to sum to one).
}
\keyword{internal}
