% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{count.each}
\alias{count.each}
\title{Count the number of instances of every unique value of a vector.}
\usage{
count.each(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector to be counted.}
}
\description{
Count the number of instances of every unique value of a vector.
}
\keyword{internal}
