% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{divide.samples}
\alias{divide.samples}
\title{Divide a given number of samples into a given number of datasets and
environments.}
\usage{
divide.samples(n.samples, n.dsets, n.envirs)
}
\arguments{
\item{n.samples}{Number of samples to divide}

\item{n.dsets}{Number of datasets to divide samples into}

\item{n.envirs}{Number of environments to divide samples into}
}
\value{
A list:
  \item{dset}{A list of vectors indicating which samples go with which
    datasets (see \code{?split}).}
  \item{env}{Like \code{dset} for environments.}
  \item{metadata}{A data frame of metadata associating samples with datasets
    and environments.}
}
\description{
Divide a given number of samples into a given number of datasets and
environments.
}
\keyword{internal}
