% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{do.fisher}
\alias{do.fisher}
\title{Wrapper around \code{fisher.test} that starts with string vectors instead of
a contingency table.}
\usage{
\method{do}{fisher}(list1, list2, background, alt = "two.sided")
}
\arguments{
\item{list1}{A vector of strings (e.g., names of significant gene hits).}

\item{list2}{A vector of strings (e.g., names of genes in a pathway).}

\item{background}{A vector of strings that \code{list1} and \code{list2} were
drawn from (e.g., names of all genes tested in an assay). Any strings in
\code{list1} or \code{list2} not in this vector will be discarded.}

\item{alt}{The alternative hypothesis (see \code{?fisher.test}).}
}
\value{
The results of \code{fisher.test} on a contingency table generated
    from \code{list1}, \code{list2}, and \code{background}, augmented with
    the additional field \code{overlap} which gives the intersection of
    \code{list1} and \code{list2}.
}
\description{
Wrapper around \code{fisher.test} that starts with string vectors instead of
a contingency table.
}
