% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy.g2s}
\alias{dummy.g2s}
\title{Make a dummy gene-to-species matrix by subsampling.}
\usage{
dummy.g2s(g2s.matrices, n, fp = 0.1, minN = 2)
}
\arguments{
\item{g2s.matrices}{A list of sparse binary matrices, one per phylum.}

\item{n}{Integer: how many genes per phylum to sample?}

\item{fp}{Double: what proportion of highest-variance genes to sample from?}

\item{minN}{Integer: minimum number of genes to return in a matrix}
}
\description{
For each matrix in the input list, find the top percentile most-variable
genes (using formula for binomial distribution), then subsample the matrix,
returning at most \code{n} genes. Fewer genes will be returned if the total
number of original genes or the number above the percentile cutoff is lower
than the desired \code{n}. Matrices with 1 or fewer genes after sampling will
be dropped.
}
\keyword{internal}
