% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy.trees}
\alias{dummy.trees}
\title{Downsample trees.}
\usage{
dummy.trees(trees, n = 50)
}
\arguments{
\item{trees}{List of trees.}

\item{n}{Number of taxa to retain (maximum)}
}
\description{
Downsample trees.
}
\keyword{internal}
