% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{equiv_test}
\alias{equiv_test}
\title{Equivalence test based on two one-sided tests.}
\usage{
equiv_test(fx, se, df, min_fx = 0.25)
}
\arguments{
\item{fx}{Estimated effect size.}

\item{se}{Estimated standard error of the effect size.}

\item{df}{Degrees of freedom.}

\item{min_fx}{Minimum effect size we care about.}
}
\value{
A p-value; reject non-equivalence if below alpha.
}
\description{
Equivalence test based on two one-sided tests.
}
\keyword{internal}
