% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{fdr.bh}
\alias{fdr.bh}
\title{Wrapper around \code{p.adjust('BH')}.}
\usage{
fdr.bh(x)
}
\arguments{
\item{x}{A vector of p-values.}
}
\value{
A vector of BH FDR values.
}
\description{
Wrapper around \code{p.adjust('BH')}.
}
\keyword{internal}
