% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{fdr.by}
\alias{fdr.by}
\title{Wrapper around \code{p.adjust('BY')}.}
\usage{
fdr.by(x)
}
\arguments{
\item{x}{A vector of p-values.}
}
\value{
A vector of BY FDR values.
}
\description{
Wrapper around \code{p.adjust('BY')}.
}
\keyword{internal}
