% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{first.element.at.depth}
\alias{first.element.at.depth}
\title{Find the first list element at a desired depth.}
\usage{
first.element.at.depth(l, n)
}
\arguments{
\item{l}{A list.}

\item{n}{A number giving a particular depth of nesting.}
}
\value{
The first element of a list
}
\description{
\code{first.element.at.depth} returns the first list element if \code{n} is
1; otherwise, the function recursively dives into nested lists in the first
list element until it reaches level \code{n}.
}
\keyword{internal}
