% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{fix.tree}
\alias{fix.tree}
\title{Fudge trees with unresolved polytomies.}
\usage{
fix.tree(phy, len = 1e-06)
}
\arguments{
\item{len}{Zero-length branches will be replaced with this fraction of the
maximum node height.}
}
\description{
\code{fix.tree} converts polytomies to dichotomies with very small branch
lengths.
}
\keyword{internal}
