% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gene.annot}
\alias{gene.annot}
\title{Annotate genes using a gene-to-function table.}
\usage{
gene.annot(x, gene.to.fxn)
}
\arguments{
\item{x}{A gene (string) or vector of genes (strings).}

\item{gene.to.fxn}{A data frame with at least "gene" and "function" as columns.}
}
\value{
A character vector of gene functions, with names equal to \code{x}.
}
\description{
Annotate genes using a gene-to-function table.
}
