% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{generate.fake.abd.meta}
\alias{generate.fake.abd.meta}
\title{Wrapper script to generate fake data and metadata.}
\usage{
generate.fake.abd.meta(n.samples = 100, n.taxa = 1000, n.envs = 3,
  n.dsets = 2, n.reads = 1e+06, env.fx.sizes = c(0, 1, -1),
  dset.fx.sizes = c(0, 0.1, -0.1), env.frac.affected = 0.25,
  dset.frac.affected = 0.1, env.sd = 0, dset.sd = 0,
  prev.dist = c(-4, 2), make.16s = FALSE, tag.length = 100,
  t.names = NULL, ...)
}
\arguments{
\item{n.samples}{Positive integer: how many samples should be simulated?}

\item{n.taxa}{Positive integer: how many taxa should be simulated?}

\item{n.envs}{Positive integer: how many environments should be simulated?}

\item{n.dsets}{Positive integer: how many datasets should be simulated?}

\item{n.reads}{Number or numeric vector giving read depth(s) per sample.}

\item{env.fx.sizes}{A numeric vector of effect sizes, one per environment.}

\item{dset.fx.sizes}{A numeric vector of effect sizes, one per dataset.}

\item{env.frac.affected}{Proportion of taxa that will be affected by an
environment.}

\item{dset.frac.affected}{Proportion of taxa that will be affected by a
dataset effect.}

\item{env.sd}{Number: when sampling environment effect sizes, use this
standard deviation.}

\item{dset.sd}{Number: when sampling dataset effect sizes, use this standard
deviation.}

\item{prev.dist}{Numeric vector of length 2 giving beta parameters for the
distribution of taxon prevalences.}

\item{make.16s}{Simulate a 16S dataset instead of a WGS-MIDAS dataset.}

\item{tag.length}{If simulating 16S data, amplicon sequence variants will be
this length.}
}
\value{
A list (comparable to what is read in by *phylogenize*):
  \item{mtx}{Simulated abundance matrix}
  \item{metadata}{Simulated metadata data frame}
}
\description{
Wrapper script to generate fake data and metadata.
}
\keyword{internal}
