% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{generate.test.pzdb}
\alias{generate.test.pzdb}
\title{Wrapper to generate a low-memory, fast-running dummy database. Make sure
\code{Matrix} is loaded.}
\usage{
generate.test.pzdb(nt = 75, ng = 50, fp = 0.1, minN = 2, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{nt}{Number of species to sample per taxon.}

\item{ng}{Number of genes to sample per taxon.}

\item{fp}{Fraction of most-variable genes to sample.}

\item{minN}{Only keep species with at least this many genes.}
}
\description{
Wrapper to generate a low-memory, fast-running dummy database. Make sure
\code{Matrix} is loaded.
}
\keyword{internal}
