% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{geommean}
\alias{geommean}
\title{Calculate geometric mean of a vector of values.}
\usage{
geommean(x)
}
\arguments{
\item{x}{Numeric vector of values.}
}
\description{
Calculate geometric mean of a vector of values.
}
\keyword{internal}
