% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-16S.R
\name{get.appspam.results}
\alias{get.appspam.results}
\title{Read in results from AppSpam}
\usage{
get.appspam.results(..., .opts = NULL)
}
\arguments{
\item{...}{Option overrides; see \code{?pz.options}.}

\item{.opts}{Internal resolved options object.}
}
\value{
A list containing a vector of hits, a vector of MIDAS ID targets, and a data
frame of the assignments as they came out of AppSpam.
}
\description{
Read and parse AppSpam \code{.jplace} output.
}
\details{
Some particularly relevant options are:
\describe{
  \item{jplace_file}{String. Path giving where to read the input .jplace file.}
}
}
