% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{get.burst.results}
\alias{get.burst.results}
\title{Read in results from BURST.}
\usage{
get.burst.results(...)
}
\value{
List containing a vector of hits, a vector of MIDAS ID targets, and a
    data frame of the assignments as they came out of BURST.
}
\description{
\code{get.burst.results} reads and parses the output of BURST to get the
best-hit MIDAS species identifier for any 16S hit. Note that the reference
16S FASTA database file must describe entries in the format: ">gene
species_or_genus_ID MIDAS_ID". Only MIDAS_ID is used so the contents of
"gene" and "species_or_genus_ID" can be arbitrary.
}
\details{
Some particularly relevant global options are:
\describe{
  \item{in_dir}{String. Path to input directory (i.e., where to look for
  input files).}
  \item{burst_outfile}{String. File name where BURST writes output which is
  then read back into \emph{phylogenize}.}
}
}
