% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{get.pheno.plotting.scales.prevalence}
\alias{get.pheno.plotting.scales.prevalence}
\title{Get phenotype plotting scales (prevalence-specific).}
\usage{
get.pheno.plotting.scales.prevalence(phenotype, trees, phenoP = 0, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{phenotype}{A named vector giving the phenotype for each taxon ID.}

\item{trees}{A list of tree objects.}

\item{phenoP}{An optional value giving the prior probability for the
environment of interest.}
}
\value{
A list of overall limits (\code{limits}), taxon-specific limits
    (\code{phy.limits}), a color scale (\code{colors}), and the zero point
    (\code{zero}).
}
\description{
Some particularly relevant options are:
\describe{
  \item{prev_color_low}{String. When graphing prevalence on a tree, this
  color is the lowest value.}
  \item{prev_color_high}{String. When graphing prevalence on a tree, this
  color is the highest value.}
}
}
\keyword{internal}
