% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-16S.R
\name{get.vsearch.results}
\alias{get.vsearch.results}
\title{Read in results from vsearch.}
\usage{
get.vsearch.results(..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}
}
\value{
List containing a vector of hits, a vector of MIDAS ID targets, and a
    data frame of the assignments as they came out of vsearch.
}
\description{
\code{get.vsearch.results} reads and parses the output of vsearch to get the
best-hit MIDAS species identifier for any 16S hit. Note that the reference
16S FASTA database file must describe entries in the format: ">gene;;
species_or_genus_ID;;MIDAS_ID". Only MIDAS_ID is used so the contents of
"gene" and "species_or_genus_ID" can be arbitrary.
}
\details{
Some particularly relevant options are:
\describe{
  \item{vsearch_outfile}{String. File name where vsearch writes output
  which is then read back into \emph{phylogenize}.}
}
}
