% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{gg.cont.tree}
\alias{gg.cont.tree}
\title{Plot continuous trait on a tree.}
\usage{
gg.cont.tree(phy, ctrait, cAnc = NULL, model = "ARD",
  cLimits = logit(c(0.025, 0.1)), n = NULL, reduced.phy = NULL,
  colors = c(low.col = "slateblue", mid.col = "black", high.col =
  "orange2"), plot = T, restrict = NULL, cName = "prevalence",
  reverse = F, ladderize = T, ...)
}
\arguments{
\item{phy}{A \code{phylo} object.}

\item{ctrait}{A named numeric vector assigning trait values to tree tips.}

\item{cAnc}{Calculated ancestry for continuous trait; if this is NULL, it is calculated.}

\item{model}{Model for calculating ancestry (see phytools::fastAnc).}

\item{cLimits}{Scale bar limits for plotting continuous trait.}

\item{n}{Character vector giving which nodes to display; if NULL, defaults to intersection of \code{phy$tip.label} and \code{names(ctrait)}.}

\item{reduced.phy}{Dichotomous tree with only the nodes in \code{n} represented; if NULL, this is calculated.}

\item{colors}{Named character vector with at least "low.col" and "high.col" and optionally "mid.col" defined, giving colors to use for plotting.}

\item{plot}{Whether to plot the tree object or just return it.}

\item{restrict}{Character vector giving which continuous trait values to plot; if NULL, all are used.}

\item{cName}{String giving the title of the plot.}

\item{reverse}{Mirror the resulting plotted tree.}

\item{ladderize}{Ladderize the plotted tree.}

\item{...}{Additional arguments passed to ggplot2.}
}
\value{
A ggtree plot of a continuous trait plotted along a tree.
}
\description{
\code{gg.cont.tree} paints a continuous trait along a tree.
}
