% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%btwn.inc\%}
\alias{\%btwn.inc\%}
\title{Test whether a value is between two other values (inclusive).}
\usage{
x \%btwn.inc\% y
}
\arguments{
\item{x}{Value(s) to test (numeric vector).}

\item{y}{Numeric vector of length 2, giving minimum and maximum values of \code{x}.}
}
\description{
Test whether a value is between two other values (inclusive).
}
\keyword{internal}
