% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%withnames\%}
\alias{\%withnames\%}
\title{Assign names to a vector.}
\usage{
x \%withnames\% y
}
\arguments{
\item{x}{Vector to have names assigned (any type).}

\item{y}{Character vector giving new names for values in \code{x}.}
}
\description{
Assign names to a vector.
}
\keyword{internal}
