% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grepv}
\alias{grepv}
\title{Abbreviation for \code{grep} with \code{value=TRUE}.}
\usage{
grepv(x, y, ...)
}
\arguments{
\item{x}{Pattern to find.}

\item{y}{Values to search for pattern \code{x}.}
}
\description{
Abbreviation for \code{grep} with \code{value=TRUE}.
}
\keyword{internal}
