% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hack.tree.labels}
\alias{hack.tree.labels}
\title{XML hack to make interactive tree diagrams.}
\usage{
hack.tree.labels(tree.obj, file, stroke.scale = 0.7, pheno = NULL,
  pheno.name = NULL, native.tooltip = FALSE, units = "", ...)
}
\arguments{
\item{tree.obj}{A ggtree representation of a tree.}

\item{file}{A filename where the final SVG output will be written.}

\item{stroke.scale}{Multiplier of stroke width in dendrogram.}

\item{pheno}{A vector with names corresponding to the tips of the tree and
values corresponding to the phenotype value at that tip.}

\item{pheno.name}{The name of the phenotype being calculated (e.g.
"prevalence").}

\item{native.tooltip}{Instead of using mouseover, use SVG tooltips (less
powerful).}

\item{units}{Postfix for the values in \code{pheno} (e.g. "%" for
percentages).}
}
\description{
This hack is very ugly but works most of the time. However, it is a good idea
to wrap it in a tryCatch so that you can fall back to a less flashy
implementation, because it relies on editing a poorly-annotated SVG file as
if it were an XML document.
}
