% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{import.pz.db}
\alias{import.pz.db}
\title{Import the data necessary for *phylogenize* analysis.}
\usage{
import.pz.db(..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}
}
\value{
A list of the data objects required to perform a *phylogenize*
    analysis, with components \code{gene.presence}, \code{trees},
    \code{taxonomy}, \code{g.mappings}, and \code{gene.to.fxn}.
}
\description{
\code{import.pz.db} decides based on resolved options which data files to
import.
}
\details{
Some particularly relevant options are:
\describe{
  \item{type_16S}{Boolean. If 16S data, TRUE, otherwise shotgun data is
    assumed. Default: FALSE}
  \item{db}{String. Which database to use ("gtdb" or "uhgp")).}
  \item{data_dir}{String. Path to directory containing the data files
  required to perform a \emph{phylogenize} analysis.}
}
}
