% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{indiv.enr}
\alias{indiv.enr}
\title{Internal function to perform an individual enrichment.}
\usage{
indiv.enr(phylum, cutoff, termset, term, data, results, qcuts = c(strong
  = 0.05), dirxn = 1, bg = NULL)
}
\description{
Internal function to perform an individual enrichment.
}
\keyword{internal}
