% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{install_data}
\alias{install_data}
\title{Copy and/or unzip downloaded data into the Phylogenize2 database directory.}
\usage{
install_data(path, force = FALSE)
}
\arguments{
\item{path}{Provide a path to a local file containing a .zip file (which will
be unzipped) or a .csv file (which will be copied).}

\item{force}{Boolean; overwrite existing files? (Default: FALSE)}
}
\description{
Copy and/or unzip downloaded data into the Phylogenize2 database directory.
}
