% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{is.dna}
\alias{is.dna}
\title{Check if a particular string is likely to be DNA.}
\usage{
is.dna(seq)
}
\arguments{
\item{seq}{String to check for illegal characters.}
}
\value{
TRUE if it contains no illegal characters, FALSE otherwise.
}
\description{
Check if a particular string is likely to be DNA.
}
\keyword{internal}
