% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kable.recolor}
\alias{kable.recolor}
\title{Function to obtain HTML colors for a particular value.}
\usage{
kable.recolor(x, direction = 1, option = "D", na_color = "#BBBBBB",
  scale_from = NULL, colors = c("#000000", "#FFFFFF"),
  limits = c(-Inf, Inf))
}
\arguments{
\item{x}{A value or vector of values to colorize.}

\item{direction}{If 1, use the color scale as given; if -1, reverse it.}

\item{na_color}{Set NA elements to this color.}

\item{scale_from}{Instead of the minimum and maximum of \code{x}, scale
values from this minimum and maximum (see \code{?rescale}).}

\item{colors}{A vector of two strings giving the low and high color,
respectively.}

\item{limits}{A vector of two numbers giving the minimum and maximum value
outside which values will be represented by the bottom or top of the
color scale, respectively.}
}
\value{
An HTML color.
}
\description{
Function to obtain HTML colors for a particular value.
}
