% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{keep.tips}
\alias{keep.tips}
\title{Keep some tips on a tree.}
\usage{
keep.tips(tree, keep)
}
\arguments{
\item{tree}{A \code{phylo} object.}

\item{keep}{A character vector of tip labels. Any tip not in this vector will
be dropped.}
}
\description{
\code{keep.tips} keeps only the set of specified tips in a tree.
}
