% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list.depth}
\alias{list.depth}
\title{Calculate the maximum depth of a list of lists.}
\usage{
list.depth(this)
}
\arguments{
\item{this}{A list, possibly containing other lists.}
}
\description{
\code{list.depth} calculates the maximum depth of a list of lists. Code is
modified from original by user "flodel" on stackoverflow: see
https://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r.
}
