% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list.even}
\alias{list.even}
\title{Make sure that a list of lists has even depth all the way down.}
\usage{
list.even(this)
}
\arguments{
\item{this}{A list, possibly containing other lists.}
}
\description{
Make sure that a list of lists has even depth all the way down.
}
