% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{lm.fx.pv}
\alias{lm.fx.pv}
\title{Wrapper around \emph{lm} that returns just the effect size and p-value.}
\usage{
lm.fx.pv(m, p, tr, coefname = "mTRUE", restrict = NULL,
  meas_err = FALSE)
}
\arguments{
\item{m}{Named numeric vector of gene presence/absences per taxon.}

\item{p}{Named numeric vector of phenotype values per taxon.}

\item{tr}{Phylogeny relating taxa (class \code{"phylo"}).}
}
\value{
Length-2 numeric vector with names \code{"Estimate"} and
    \code{"p.value"}. If there is an error in \code{phylolm}, the values of
    this vector will be \code{c(NA, NA)}.
}
\description{
Wrapper around \emph{lm} that returns just the effect size and p-value.
}
