% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logistic}
\alias{logistic}
\title{Calculate inverse-logit of a value or vector of values.}
\usage{
logistic(x)
}
\arguments{
\item{x}{Numeric value, or numeric vector of numeric values.}
}
\description{
Calculate inverse-logit of a value or vector of values.
}
