% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logit}
\alias{logit}
\title{Calculate logit of a value or vector of values.}
\usage{
logit(x)
}
\arguments{
\item{x}{Numeric value, or numeric vector of numeric values.}
}
\description{
Calculate logit of a value or vector of values.
}
