% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{make.pos.sig}
\alias{make.pos.sig}
\title{Get vectors of significant genes with positive effect sizes.}
\usage{
make.pos.sig(sigs, signs, cut = "strong")
}
\arguments{
\item{sigs}{Output of \code{make.sigs}}

\item{signs}{Output of \code{make.signs}}

\item{cut}{String giving named significance level to use.}
}
\value{
List (per phylum) of string vectors of positive significant hits.
}
\description{
Get vectors of significant genes with positive effect sizes.
}
