% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{make.signs}
\alias{make.signs}
\title{Get effect sizes of genes from result tables.}
\usage{
make.signs(results)
}
\arguments{
\item{results}{List of result matrices with two rows (effect size and
p-value) and one column per gene tested.}
}
\value{
List (per phylum) of numeric vectors of signs of hits.
}
\description{
Get effect sizes of genes from result tables.
}
