% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{make.sim}
\alias{make.sim}
\title{Construct a simulation of microbial sequencing data.}
\usage{
make.sim(n.taxa, n.samples, avg.reads, prev.dist = c(-4, 2),
  effect.size.mtx = 0, alphas = NULL)
}
\arguments{
\item{n.taxa}{Integer; number of taxa to simulate.}

\item{n.samples}{Integer; number of samples to simulate.}

\item{avg.reads}{Number or numeric vector giving read depth(s) per sample.}

\item{prev.dist}{Two-element numeric vector giving beta distribution
parameters for the prevalence distribution.}

\item{effect.size.mtx}{An optional numeric matrix of true effect sizes (in
logit space).}

\item{alphas}{Optionally, a numeric vector of Dirichlet alpha parameters to
use for Dirichlet-Multinomial simulation.}
}
\value{
A list:
  \item{sim}{A simulated abundance matrix}
  \item{com}{Multinomial distribution used to simulate abundances}
  \item{prev}{Distribution of taxon prevalences.}
}
\description{
Construct a simulation of microbial sequencing data.
}
\keyword{internal}
