% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{master.fx.matrix}
\alias{master.fx.matrix}
\title{Construct a master effect size matrix given environmental and dataset
effects (wraps \code{master.fx.matrix}).}
\usage{
master.fx.matrix(divided, taxa, env.fx, env.taxa, dset.fx, dset.taxa)
}
\arguments{
\item{divided}{Output of \code{divide.samples}.}

\item{taxa}{Vector of strings representing taxon names.}

\item{env.fx}{Named numeric vector giving effect sizes per environment.}

\item{env.taxa}{List of vectors of strings giving the subset of taxa to be
affected in each distinct environment effect.}

\item{dset.fx}{Named numeric vector giving effect sizes per dataset.}

\item{dset.taxa}{List of vectors of strings giving the subset of taxa to be
affected in each distinct dataset effect.}
}
\value{
An effect size matrix incorporating both environmental and dataset
    effects.
}
\description{
Construct a master effect size matrix given environmental and dataset
effects (wraps \code{master.fx.matrix}).
}
\keyword{internal}
