% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{matrix.plm}
\alias{matrix.plm}
\title{Perform phylogenetic (or linear) modeling for a single taxon.}
\usage{
matrix.plm(
  tree,
  mtx,
  pheno,
  method = phylolm.fx.pv,
  restrict.taxa = NULL,
  restrict.ff = NULL,
  ...,
  .opts = NULL
)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{tree}{A tree relating taxa within a taxon.}

\item{mtx}{Gene presence/absence matrix.}

\item{pheno}{Named numeric vector giving phenotype values per species.}

\item{method}{A function that returns a length-2 numeric vector of
effect-size and p-value (see, e.g., \code{phylolm.fx.pv} or
\code{lm.fx.pv}).}

\item{restrict.taxa}{Optionally, a character vector giving a subset of
taxa to test.}

\item{restrict.ff}{Optionally, a character vector giving a subset of
genes to test.}
}
\value{
Matrix of p-values (row 1) and effect-sizes (row 2) per gene
    (columns).
}
\description{
Perform phylogenetic (or linear) modeling for a single taxon.
}
