% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{maybeParApply}
\alias{maybeParApply}
\title{A wrapper around \code{apply} and \code{parApply} that allows them to be
called with a single syntax.}
\usage{
maybeParApply(mtx, margin, fun, cl = NULL, ...)
}
\arguments{
\item{mtx}{Matrix to apply a function over.}

\item{margin}{Margin of matrix to iterate over.}

\item{fun}{Function to apply over matrix.}

\item{cl}{If NULL, \code{apply} will be called; otherwise, should be the
object returned by \code{makeCluster}.}
}
\description{
A wrapper around \code{apply} and \code{parApply} that allows them to be
called with a single syntax.
}
\keyword{internal}
