% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{min.nonzero}
\alias{min.nonzero}
\title{Find the minimum value of a vector that is still greater than zero.}
\usage{
\method{min}{nonzero}(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Find the minimum value of a vector that is still greater than zero.
}
\keyword{internal}
