% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{multi.enrich}
\alias{multi.enrich}
\title{Given lists of significant genes (at different thresholds), effect sizes, and
gene set mappings, assemble a tbl of results.}
\usage{
multi.enrich(sigs, signs, mappings, dirxn = 1)
}
\arguments{
\item{sigs}{List giving, per phylum (outer) and per significance cutoff
(inner), significant hits to test for enrichment.}

\item{signs}{List giving, per phylum, signs of all gene effect sizes.}

\item{mappings}{List of data.frames giving gene-to-gene-set mappings.}

\item{dirxn}{Count only genes with this effect sign as significant.}
}
\value{
A tbl giving Fisher's test p-values, q-values, effect sizes, and overlaps.
}
\description{
Given lists of significant genes (at different thresholds), effect sizes, and
gene set mappings, assemble a tbl of results.
}
