% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{NameFaker}
\alias{NameFaker}
\alias{get.species.names}
\alias{get.sample.names}
\title{Generate fake names for species or samples.}
\usage{
get.species.names(n.species)

get.sample.names(n.samples)
}
\arguments{
\item{n.species}{Number of species.}

\item{n.samples}{Number of samples.}
}
\value{
A vector of fake names.
}
\description{
Generate fake names for species or samples.
}
\keyword{internal}
