% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{optimize.b.wrapper}
\alias{optimize.b.wrapper}
\title{Based on a real presence/absence matrix, optimize the value of the
regularization parameter $b$.}
\usage{
optimize.b.wrapper(real.mtx, real.ids, which.real.env = 2,
  which.shape = 1, prior = 0.002, effect.size = 2, bounds = c(0.01,
  5), add.pc = FALSE, tol = prior * 0.005, a = 0.05,
  verbose = FALSE, optim.fxn = b.scorer, pos.prop = 0.5, ...)
}
\arguments{
\item{real.mtx}{Presence/absence matrix.}

\item{real.ids}{A factor assigning environment labels to samples.}

\item{which.real.env}{A numeric or string giving the environment to calculate
specificity within.}

\item{which.shape}{A numeric or string giving the environment in which the
prevalence distribution should be fit.}

\item{prior}{Prior probability of encountering environment \code{which.real.env}.}

\item{effect.size}{Effect size to be simulated for "true" positives.}

\item{bounds}{Bounds for optimizing $b$.}

\item{add.pc}{Boolean giving whether to add a pseudocount when calculating
prevalences and specificities.}

\item{tol}{Numeric: values within \code{tol} of the prior will be considered
to be shrunk back to the prior completely.}

\item{a}{See ?\code{b.scorer}.}

\item{verbose}{Boolean: print debugging information?}

\item{optim.fxn}{Function summarizing the results of
\code{score.regularization} into one metric to be optimized.}

\item{pos.prob}{Fraction of real effects that should be positive.}
}
\value{
The output of \code{stats::optimize}.
}
\description{
Based on a real presence/absence matrix, optimize the value of the
regularization parameter $b$.
}
\keyword{internal}
