% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{output.enr.table}
\alias{output.enr.table}
\title{Make a pretty enrichment table.}
\usage{
output.enr.table(enr.table)
}
\arguments{
\item{enr.table}{Input enrichment table.}
}
\value{
Mutated enrichment table with better-labeled columns and significance
    coloring.
}
\description{
Make a pretty enrichment table.
}
