% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pblapply.across.names}
\alias{pblapply.across.names}
\title{Apply a function to a vector of names, with the returned list having those
names (progress bar).}
\usage{
pblapply.across.names(X, FUN, ...)
}
\arguments{
\item{X}{Vector of names.}

\item{FUN}{A function to apply to the names in \code{X} (typically using them
as list indices).}
}
\value{
A list of the results of applying \code{FUN} to \code{X}, with
    \code{X} as the list elements' names.
}
\description{
Apply a function to a vector of names, with the returned list having those
names (progress bar).
}
\keyword{internal}
