% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbmcapply}
\alias{pbmcapply}
\title{Apply a function over a margin of a matrix in parallel (with progress bar).}
\usage{
pbmcapply(X, MARGIN, FUN, mc.cores = 10, simplify = TRUE, ...)
}
\arguments{
\item{X}{A matrix.}

\item{MARGIN}{A number specifying whether to apply over rows (1) or columns
(2).}

\item{FUN}{A function to be applied.}

\item{mc.cores}{Number of cores to use.}

\item{simplify}{Whether to simplify the results using \code{simplify2array}.}
}
\description{
Apply a function over a margin of a matrix in parallel (with progress bar).
}
\keyword{internal}
