% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{phylolm.subset}
\alias{phylolm.subset}
\title{Wrapper around \code{phylolm} that automatically discards taxa that are
absent from the tree tip labels, phenotype labels, or gene presence/absence
labels.}
\usage{
phylolm.subset(p, m, phy)
}
\arguments{
\item{p}{Named numeric vector of phenotype values per taxon.}

\item{m}{Named numeric vector of gene presence/absences per taxon.}

\item{phy}{Phylogeny relating taxa (class \code{"phylo"}).}
}
\value{
Output of \code{phylolm}.
}
\description{
Wrapper around \code{phylolm} that automatically discards taxa that are
absent from the tree tip labels, phenotype labels, or gene presence/absence
labels.
}
\keyword{internal}
