% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{plot.phenotype.trees}
\alias{plot.phenotype.trees}
\title{Plot a phenotype along a list of trees.}
\usage{
\method{plot}{phenotype.trees}(phenotype, trees, taxonomy, scale, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{phenotype}{A named vector with the phenotype values for each taxon.}

\item{trees}{A list of trees.}

\item{taxonomy}{A dataframe with the taxonomix information.}

\item{scale}{A list returned from \code{get.pheno.plotting.scales}.}
}
\value{
A list of ggtree objects in which the phenotype has been plotted
    across each tree in \code{trees}.
}
\description{
Some particularly relevant options are:
\describe{
  \item{which_phenotype}{String. Which phenotype to calculate ("prevalence"
  or "specificity").}
}
}
