% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{process.16s}
\alias{process.16s}
\title{Map denoised 16S sequence variants to MIDAS IDs using vsearch.}
\usage{
process.16s(abd.meta, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{mtx}{A presence/absence or abundance matrix, with row names equal to
amplicon sequence variant DNA sequences.}
}
\value{
none
}
\description{
\code{process.16s} is a wrapper for other functions that: output sequence
variants to a file; map them using vsearch against a reference database of
16S sequences; then return a list of abundance and metadata values where the
rows of the abundance matrix are now MIDAS IDs.
}
\details{
Some particularly relevant options are:
\describe{
  \item{vsearch_infile}{String. File name of the sequences written to disk
    and then read into vsearch.}
}
}
