% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{pz.error}
\alias{pz.error}
\title{Throw an error and optionally log it in errmsg.txt.}
\usage{
pz.error(errtext, ...)
}
\arguments{
\item{errtext}{String: error message text.}
}
\description{
Some particularly relevant options are:
\describe{
  \item{error_to_file}{Boolean. Should pz.error, pz.warning, and pz.message
  output to an error message file?}
}
}
