% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{pz.warning}
\alias{pz.warning}
\title{Report a warning and optionally log it in errmsg.txt.}
\usage{
pz.warning(msgtext, ...)
}
\arguments{
\item{errtext}{String: warning text.}
}
\description{
Some particularly relevant options are:
\describe{
  \item{error_to_file}{Boolean. Should pz.error, pz.warning, and pz.message
  output to an error message file?}
}
}
