% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{qvals}
\alias{qvals}
\title{Wrapper around \code{qvalue} and \code{p.adjust} that extracts only q-values.
If there is an error in `qvalue()` with default parameters, this function
will automatically fall back to a Benjamini-Hochberg-style correction (by
setting lambda to zero), using p.adjust if that still doesn't work, and
finally returning a vector of NAs if all else fails.}
\usage{
qvals(x, ...)
}
\arguments{
\item{x}{A vector of p-values.}

\item{...}{Extra parameters to override defaults, especially `fdr_method` (which can be "BH", "BY", or "qvalue).}
}
\value{
A vector of q-values.
}
\description{
Wrapper around \code{qvalue} and \code{p.adjust} that extracts only q-values.
If there is an error in `qvalue()` with default parameters, this function
will automatically fall back to a Benjamini-Hochberg-style correction (by
setting lambda to zero), using p.adjust if that still doesn't work, and
finally returning a vector of NAs if all else fails.
}
\keyword{internal}
