% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{read.abd.metadata.biom}
\alias{read.abd.metadata.biom}
\title{Read abundance matrix and metadata (BIOM)}
\usage{
read.abd.metadata.biom(..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}
}
\value{
A list with components \code{mtx} (matrix of abundances) and
    \code{metadata} (data frame of metadata).
}
\description{
Read in taxon-by-sample matrix of abundances and metadata (sample
annotations) from a single BIOM-formatted file.
}
\details{
This function resolves options from \code{pz.options()} defaults
    plus any overrides supplied through \code{...}. Some particularly relevant
    options are:

\describe{
  \item{biom_file}{String. Name of BIOM abundance-and-metadata file. Default: "test.biom"}
}
}
\keyword{internal}
