% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{read.abd.metadata}
\alias{read.abd.metadata}
\title{Read in abundance and metadata file(s).}
\usage{
read.abd.metadata(..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}
}
\value{
A list with components \code{mtx} and \code{metadata}, corresponding
    to a sparse binary presence/absence matrix (see \code{Matrix} package)
    and a metadata data frame.
}
\description{
Read in abundance and metadata, either as one BIOM-format file or as two
tab-delimited files.
}
\details{
This function resolves options from \code{pz.options()} defaults
    plus any overrides supplied through \code{...}. Some particularly relevant
    options are:

\describe{
  \item{env_column}{String. Name of column in metadata file containing the
    environment annotations.}
  \item{dset_column}{String. Name of column in metadata file containing the
    dataset annotations.}
  \item{input_format}{String. Whether to look for tabular or BIOM-formatted
    data ("tabular" or "biom").}
  \item{type}{String. Type of data to use, either "midas" (shotgun) or "16S"
    (amplicon).}
}
}
