% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{read.abd.metadata.tabular}
\alias{read.abd.metadata.tabular}
\title{Read abundance matrix and metadata (tabular)}
\usage{
read.abd.metadata.tabular(..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}
}
\value{
A list with components \code{mtx} (matrix of abundances) and
    \code{metadata} (data frame of metadata).
}
\description{
Read in taxon-by-sample matrix of abundances and metadata (sample
annotations) from two tab-delimited files.
}
\details{
Some particularly relevant options are:
\describe{
  \item{dset_column}{Name of metadata column containing dataset annotations.}
}
}
\keyword{internal}
