% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{rem_unobs_clades}
\alias{rem_unobs_clades}
\title{Remove clades above a certain size with no observed tips.}
\usage{
rem_unobs_clades(phy, observed, pct = 0.1)
}
\arguments{
\item{phy}{A phylo object.}

\item{observed}{A character vector giving all the observed tips. May contain
tips not in the tree, which will be ignored.}

\item{pct}{Double; a cutoff. For example, \code{pct}=0.1 means that any
unobserved clade containing at least 10% of the total number of tips in
the tree will be dropped.}
}
\value{
A new, trimmed phylo object.
}
\description{
Remove clades above a certain size with no observed tips.
}
\keyword{internal}
