% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{remove.allzero.abundances}
\alias{remove.allzero.abundances}
\title{Remove rows and columns of a matrix that are all zero.}
\usage{
remove.allzero.abundances(abd.mtx, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{abd.mtx}{A matrix of abundance values (double or logical).}
}
\value{
A matrix of abundance values (double), with all-zero columns and rows
    removed.
}
\description{
\code{remove.allzero.abundances} removes all rows and columns of a matrix
where every observation is zero, starting with columns and then proceeding to
rows.
}
