% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{render.report}
\alias{render.report}
\title{Run *phylogenize* start to finish.}
\usage{
render.report(output_file = "report_output.html",
  report_input = "phylogenize-report.Rmd", do_cache = TRUE,
  reset_after = TRUE, ...)
}
\arguments{
\item{output_file}{Path giving what to name the resulting HTML file.}

\item{report_input}{Optionally override which notebook to knit (useful for
testing).}

\item{do_cache}{Turn on or off Rmarkdown's caching.}

\item{reset_after}{Reset global options to package defaults after running?}

\item{...}{Parameters to override defaults.}
}
\description{
Run *phylogenize* start to finish.
}
