% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{rep.col}
\alias{rep.col}
\title{Convenience function to repeat a vector as columns.}
\usage{
\method{rep}{col}(vec, nc)
}
\arguments{
\item{vec}{Vector to repeat.}

\item{nc}{Number of times to repeat vector.}
}
\value{
A matrix consisting of \code{nc} repeats of the column-vector
    \code{vec}.
}
\description{
Convenience function to repeat a vector as columns.
}
\keyword{internal}
