% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{results.report}
\alias{results.report}
\title{Create a summary table giving how many tests were significant.}
\usage{
results.report(results, sigs, signs)
}
\arguments{
\item{results}{List of result matrices, one per phylum.}

\item{sigs}{The output of \code{make.sigs}.}

\item{signs}{The output of \code{make.signs}.}
}
\value{
A table with the number of positive significant results per phylum at
    each significance level in \code{sigs}.
}
\description{
Create a summary table giving how many tests were significant.
}
