% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{retain.observed.taxa}
\alias{retain.observed.taxa}
\title{Modify trees to retain only observed taxa (for use with specificity only).}
\usage{
retain.observed.taxa(trees, phenotype, phenoP, mapped.observed)
}
\arguments{
\item{trees}{A list of tree objects.}

\item{phenotype}{A named vector giving the phenotype for each taxon ID.}

\item{phenoP}{The prior probability of the environment of interest.}

\item{mapped.observed}{A character vector giving which tips to retain.}
}
\value{
An updated list of tree objects.
}
\description{
Modify trees to retain only observed taxa (for use with specificity only).
}
