% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{rnd.fx.pv}
\alias{rnd.fx.pv}
\title{Wrapper to return random effect sizes and p-values; useful for testing.}
\usage{
rnd.fx.pv(m, p, tr, coefname = "mTRUE", restrict = NULL,
  meas_err = FALSE)
}
\arguments{
\item{m}{Named numeric vector of gene presence/absences per taxon.}

\item{p}{Named numeric vector of phenotype values per taxon.}

\item{tr}{Phylogeny relating taxa (class \code{"phylo"}).}

\item{coefname}{Which coefficient from the phylolm to return (meaningless
here)?}

\item{restrict}{If not NULL, a character vector of taxa to consider
(meaningless here).}
}
\value{
Length-2 numeric vector with names \code{"Estimate"} and
    \code{"p.value"}, with distributions N(0,1) and U(0,1), respectively.
}
\description{
Wrapper to return random effect sizes and p-values; useful for testing.
}
\keyword{internal}
