% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-16S.R
\name{run.appspam}
\alias{run.appspam}
\title{Run AppSpam analysis on a FASTA file of sequences}
\usage{
run.appspam(..., .opts = NULL)
}
\arguments{
\item{...}{Option overrides; see \code{?pz.options}.}

\item{.opts}{Internal resolved options object.}
}
\value{
Returns TRUE unless an error is thrown.
}
\description{
Run AppSpam for phylogenetic placement of named 16S sequence variants.
}
\details{
Some particularly relevant options are:
\describe{
  \item{appspam_path}{String. Path to appspam binary.}
  \item{aln_path_16s}{String. Path to file containing aligned 16S sequences.}
  \item{tree_path_16s}{String. Path to tree for phylogenetic placement.}
  \item{named_asv_file}{String. Path to FASTA file containing named ASVs.}
  \item{jplace_file}{String. Path giving where to store output .jplace file.}
}
}
