% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{run.burst}
\alias{run.burst}
\title{Run vsearch analysis on a FASTA file of sequences.}
\usage{
\method{run}{burst}(...)
}
\value{
Returns TRUE unless an error is thrown.
}
\description{
\code{run.burst} runs the optimal sequence aligner vsearch
(doi.org/10.7717/peerj.2584) on a FASTA file, typically one generated by
\code{prepare.vsearch.input}. (In theory, by changing the \code{burst_bin}
option, any aligner could be used provided it accepts the same command-line
options and returns the same formatted output as vsearch.)
}
\details{
Some particularly relevant global options are:
\describe{
  \item{in_dir}{String. Path to input directory (i.e., where to look for
input files).}
  \item{burst_infile}{String. File name of the sequences to be read into
vsearch.}
  \item{vsearch_outfile}{String. File name where vsearch writes output which is
then read back into \emph{phylogenize}.}
  \item{burst_dir}{String. Path where the binary of vsearch is found.}
  \item{burst_bin}{String. File name of the binary of vsearch.}
  \item{vsearch_16sfile}{String. Path to the 16S FASTA database that maps back
to MIDAS species.}
  \item{data_dir}{String. Path to directory containing the data files
required to perform a \emph{phylogenize} analysis. }
}
}
