% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{sanity.check.abundance}
\alias{sanity.check.abundance}
\title{Sanity-check abundance data}
\usage{
sanity.check.abundance(abd.mtx, ...)
}
\arguments{
\item{abd.mtx}{A matrix or Matrix of abundance or presence values (double or
logical).}
}
\value{
Always returns TRUE, but will throw errors if the
    abundance data is the wrong type or class.
}
\description{
\code{sanity.check.abundance} is used to make sure that the abundance matrix
satisfies the requirements specified by the \emph{phylogenize} application.
}
