% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-process.R
\name{sanity.check.metadata}
\alias{sanity.check.metadata}
\title{Check that dataset, environment, and sample columns all present}
\usage{
sanity.check.metadata(metadata, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{metadata}{A data frame giving sample annotations.}
}
\value{
Always returns TRUE, but will throw errors if the metadata does not
  match specifications.
}
\description{
\code{sanity.check.abundance} is used to make sure that the metadata data
frame satisfies the requirements specified by the \emph{phylogenize}
application.
}
\details{
Some particularly relevant options are:
\describe{
  \item{env_column}{Name of metadata column containing environment
    annotations.}
  \item{dset_column}{Name of metadata column containing dataset
    annotations.}
}
}
