% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_data_internal}
\alias{set_data_internal}
\title{Set data directory to internal}
\usage{
set_data_internal(fail = FALSE)
}
\arguments{
\item{fail}{Boolean. If TRUE, set_data_internal will not attempt to download
and install data from Figshare if it is missing.}
}
\description{
Set data directory to internal
}
