% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{simple.prevalence}
\alias{simple.prevalence}
\title{Calculate prevalence (with additive smoothing) for taxa given an abundance matrix.}
\usage{
simple.prevalence(mtx)
}
\arguments{
\item{mtx}{Abundance matrix (rows: taxa; columns: samples).}
}
\value{
Prevalence with additive smoothing.
}
\description{
Calculate prevalence (with additive smoothing) for taxa given an abundance matrix.
}
\keyword{internal}
