% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{simulate.binom.mtx}
\alias{simulate.binom.mtx}
\title{Simulate a presence/absence matrix.}
\usage{
\method{simulate}{binom.mtx}(effect.size = 2,
  baseline.distro = c(shape1 = 0.66, shape2 = 2.62), samples = c(H =
  38, D = 13), which.env = "D", taxa = 2000, tpr = 0.25,
  sign.pos.prob = 0.5)
}
\arguments{
\item{effect.size}{A number giving the shift in logit-prevalence between
environments for simulated true positives.}

\item{baseline.distro}{A two-element numeric vector of beta distribution
parameters, giving the distribution of simulated prevalences.}

\item{samples}{Named integer vector giving the number of samples per environment.}

\item{which.env}{String; gives the environment in which an effect will be simulated.}

\item{taxa}{How many taxa to simulate?}

\item{tpr}{Fraction of taxa that should be simulated as true positives.}

\item{sign.pos.prob}{Fraction of effects that should be positive (vs. negative)?}
}
\value{
A list. \code{mtx}: a simulated matrix; \code{pT}: true prevalences;
    \code{pTbs}: true prevalences after adding in effects; \code{fx}:
    effects; \code{ids}: mapping of samples to environments;
    \code{input.params}: input parameters used to call
    \code{simulate.binom.mtx}.
}
\description{
Simulate a presence/absence matrix.
}
\keyword{internal}
