% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{single.cluster.plot}
\alias{single.cluster.plot}
\title{Make a hybrid tree-heatmap plot showing the taxon distribution of significant
hits.}
\usage{
single.cluster.plot(
  gene.presence,
  sig.genes,
  tree,
  plotted.tree,
  taxon,
  verbose = FALSE,
  ...,
  .opts = NULL
)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{gene.presence}{Gene presence/absence matrix.}

\item{sig.genes}{Character vector of the significant genes.}

\item{tree}{A tree object.}

\item{plotted.tree}{A ggtree plot of \code{tree}.}

\item{taxon}{Name of the taxon represented by \code{tree}}

\item{verbose}{Whether to report debugging information (boolean).}
}
\value{
A faceted ggplot object.
}
\description{
Some particularly relevant options are:
\describe{
  \item{which_phenotype}{String. Which phenotype to calculate ("prevalence"
  or "specificity").}
  \item{gene_color_absent}{String. When graphing gene presence/absence,
  this color indicates absence.}
  \item{gene_color_present}{String. When graphing gene presence/absence, this
  color indicates presence.}
}
}
