% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{small.merge}
\alias{small.merge}
\title{Merge two vectors/matrices by (row) names, returning a matrix with row names.}
\usage{
small.merge(x, y, ...)
}
\arguments{
\item{x}{First vector or matrix.}

\item{y}{Second vector or matrix.}
}
\value{
A merged matrix with row names.
}
\description{
Merge two vectors/matrices by (row) names, returning a matrix with row names.
}
\keyword{internal}
