% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sparseMelt}
\alias{sparseMelt}
\title{Helper function to "melt" a sparse matrix into a long format.}
\usage{
sparseMelt(mtx)
}
\arguments{
\item{mtx}{An object of class \code{TsparseMatrix}. @return A data frame with
the data in \code{mtx} represented in "long" (vs. "wide") format.}
}
\description{
Helper function to "melt" a sparse matrix into a long format.
}
\keyword{internal}
