% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{style.parse}
\alias{style.parse}
\title{Helper function to parse SVG styles.}
\usage{
style.parse(str)
}
\arguments{
\item{str}{A style string to parse.}
}
\value{
A named vector of style attributes.
}
\description{
Helper function to parse SVG styles.
}
\keyword{internal}
