% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{sum.nonunique.burst}
\alias{sum.nonunique.burst}
\title{Sum non-unique rows after BURST mapping.}
\usage{
\method{sum}{nonunique.burst}(burst, mtx, ...)
}
\arguments{
\item{burst}{A list obtained by running \code{get.burst.results}.}

\item{mtx}{A presence/absence or abundance matrix, with row names equal to
amplicon sequence variant DNA sequences.}
}
\value{
A new matrix with MIDAS IDs as rows.
}
\description{
\code{sum.nonunique.burst} takes BURST results and an abundance or
presence/absence matrix, drops any rows that mapped to multiple MIDAS IDs
(i.e. that couldn't confidently be assigned to a MIDAS species),
then sums any rows that mapped to the same MIDAS ID.
}
\details{
Some particularly relevant global options are:
\describe{
  \item{out_dir}{String. Path to output directory. Default: "output"}
  \item{in_dir}{String. Path to input directory (i.e., where to look for
  input files).}
  \item{data_dir}{String. Path to directory containing the data files
  required to perform a \emph{phylogenize} analysis. Default: on package
  load, this default is set to the result of \code{system.file("extdata",
  package="phylogenize")}.}
}
}
