% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-16S.R
\name{sum.nonunique.vsearch}
\alias{sum.nonunique.vsearch}
\title{Sum non-unique rows after vsearch mapping.}
\usage{
\method{sum}{nonunique.vsearch}(vsearch, mtx, ...)
}
\arguments{
\item{vsearch}{A list obtained by running \code{get.vsearch.results}.}

\item{mtx}{A presence/absence or abundance matrix, with row names equal to
amplicon sequence variant DNA sequences.}
}
\value{
A new matrix with MIDAS IDs as rows.
}
\description{
\code{sum.nonunique.vsearch} takes vsearch results and an abundance or
presence/absence matrix, drops any rows that mapped to multiple MIDAS IDs
(i.e. that couldn't confidently be assigned to a MIDAS species),
then sums any rows that mapped to the same MIDAS ID.
}
\details{
Some particularly relevant global options are:
\describe{
  \item{min_frac_16s}{Numeric. Should be between 0.5 and 1. Only keep ASVs
  where at least this fraction of assignments are to the same species. 
  Allows some tolerance for mislabeled or phylogenetically-inconsistent
  16S sequences in the database. Default: 0.8}
  \item{out_dir}{String. Path to output directory. Default: "output"}
  \item{data_dir}{String. Path to directory containing the data files
  required to perform a \emph{phylogenize} analysis. Default: on package
  load, this default is set to the result of \code{system.file("extdata",
  package="phylogenize")}.}
}
}
