% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tax.annot}
\alias{tax.annot}
\title{Annotate taxa using a taxonomy table.}
\usage{
tax.annot(tns, taxonomy)
}
\arguments{
\item{tns}{A vector of taxon IDs.}

\item{taxonomy}{A data frame with at least "cluster" and "species" columns;
"cluster" is used to match the identifiers in \code{tns}.}
}
\value{
A character vector of species names.
}
\description{
Annotate taxa using a taxonomy table.
}
