% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{tbl.result.qvs}
\alias{tbl.result.qvs}
\title{Get q-values for results in tbl format.}
\usage{
\method{tbl}{result.qvs}(results, method = qvals, ...)
}
\arguments{
\item{results}{A tbl of results with columns for "phylum" and "p.value".}

\item{method}{A function: method for obtaining q-values from p-values.}

\item{...}{Additional parameters to pass to \code{method}.}
}
\value{
A tbl with an additional "q.value" column.
}
\description{
Get q-values for results in tbl format.
}
