% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tipToRoot}
\alias{tipToRoot}
\title{Get tip-to-root distance.}
\usage{
tipToRoot(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
A vector of root-to-tip distances.
}
\description{
\code{tipToRoot} returns all tip-to-root distances.
}
\keyword{internal}
