% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree.to.dist}
\alias{tree.to.dist}
\title{Get tip-to-tip distances.}
\usage{
tree.to.dist(tree)
}
\arguments{
\item{tree}{A \code{phylo} object.}
}
\value{
A \code{dist} object representing tip-to-tip distances.
}
\description{
\code{tree.to.dist} returns all tip-to-tip distances in distance matrix
format.
}
\keyword{internal}
