% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{truncated}
\alias{truncated}
\title{Truncate a vector with a particular lower and upper limit.}
\usage{
truncated(x, lim = c(logit(0.001), logit(0.25)))
}
\arguments{
\item{x}{Vector to truncate.}

\item{lim}{Two-element numeric vector giving the lower and upper limits.}
}
\value{
A vector where elements below (or above) the lower (or upper) limit
    have been replaced with that limit.
}
\description{
Truncate a vector with a particular lower and upper limit.
}
\keyword{internal}
