% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{write.test.biom}
\alias{write.test.biom}
\title{Write and then read in tabular data.}
\usage{
write.test.biom(abd.meta, overwrite = FALSE, ..., .opts = NULL)
}
\arguments{
\item{.opts}{Internal resolved options object.}

\item{abd.meta}{List containing a matrix of abundance values, \code{mtx}, and
a data frame of metadata, \code{metadata}.}

\item{overwrite}{Boolean: if a BIOM file exists at the location specified,
overwrite it?}
}
\description{
Some options that may be helpful include:
\describe{
  \item{biom_file}{String. Name of BIOM abundance-and-metadata file (in this
  case, to write to disk).}
  \item{biom_dir}{String. Path to BIOM command-line executables. These are
  necessary to write the BIOM file and add the metadata.}
}
}
\keyword{internal}
