% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{write.test.tabular}
\alias{write.test.tabular}
\title{Write and then read in tabular data.}
\usage{
write.test.tabular(abd.meta, abdfile = "test-abundance.tab",
  metafile = "test-metadata.tab", prep = TRUE, ...)
}
\arguments{
\item{abd.meta}{List containing a matrix of abundance values, \code{mtx}, and
a data frame of metadata, \code{metadata}.}

\item{abdfile}{String: write abundance matrix to this file.}

\item{metafile}{String: write metadata data frame to this file.}

\item{prep}{Boolean: convert matrix to a data frame and add a row name title.}
}
\value{
Return value of \code{base::write.table}.
}
\description{
Write and then read in tabular data.
}
\keyword{internal}
