% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zipData}
\alias{zipData}
\title{``Zip'' a list of data together with its names.}
\usage{
zipData(iterover)
}
\arguments{
\item{iterover}{List to iterate over.}
}
\value{
A list with the same number of elements as \code{iterover}. Each list
    element is itself a list with the slots "name" and "data". For the ith
    element of the returned list, "name" contains \code{names(iterover)[i]}
    and "data" contains \code{iterover[[i]]}.
}
\description{
This function takes a list and constructs a new structured list with fields
"name" and "data". This is typically called by \code{zipLapply} or
\code{zipSapply}. The purpose is to allow the user to iterate over a list
using lapply- or sapply-like syntax while still having access to the names of
the list elements.
}
\keyword{internal}
