% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zipSapply}
\alias{zipSapply}
\title{A wrapper for lapply that allows access to list element names, with
simplification at the end to an array.}
\usage{
zipSapply(iterover, fxn, ...)
}
\arguments{
\item{iterover}{List to iterate over. This list will be transformed by \code{zipData}.}

\item{fxn}{A function to apply to each element \code{x} of
\code{zipData(iterover)}, where the name is accessible as \code{x$name}
and the original list element is accessible as \code{x$data}.}
}
\description{
A wrapper for lapply that allows access to list element names, with
simplification at the end to an array.
}
\keyword{internal}
