% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{adjust.db}
\alias{adjust.db}
\title{Clean up imported database.}
\usage{
adjust.db(pz.db, abd.meta, ...)
}
\arguments{
\item{pz.db}{A database (typically obtained with \code{import.pz.db}).}

\item{abd.meta}{A list consisting of a taxon abundance matrix and the
metadata.}
}
\value{
An updated database.
}
\description{
\code{adjust.db} removes any phyla with fewer than \code{opts('treemin')}
representatives, removes tips from trees that were not observed in the data,
and resolves polytomies. It also adds a couple of variables into the database
that give lists of taxa per tree and the total number of remaining phyla.
}
