% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{check.process.metadata}
\alias{check.process.metadata}
\title{Check and process metadata}
\usage{
check.process.metadata(metadata, ...)
}
\arguments{
\item{metadata}{A data frame of metadata with environment, dataset, and
sample columns corresponding to those in the global options (see
\?pz.options).}
}
\value{
A data frame of metadata, with environment and
    dataset columns converted to factors.
}
\description{
\code{check.process.metadata} is used to make sure that the metadata
satisfies the requirements specified by the global options and to make sure
that the metadata are of the correct type.
}
\details{
Some particularly relevant global options are:
\describe{
  \item{env_column}{Name of metadata column containing environment annotations.}
  \item{dset_column}{Name of metadata column containing dataset annotations.}
  \item{single_dset}{Boolean. If true, will assume that all samples come from
  a single dataset called \code{dset1} no matter what, if anything, is in
  \code{dset_column}.}
}
}
