% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{do.clust.plot}
\alias{do.clust.plot}
\title{Make a hybrid tree-heatmap plot showing the taxon distribution of significant
hits.}
\usage{
\method{do}{clust.plot}(gene.presence, sig.genes, tree, plotted.tree,
  phylum, verbose = FALSE, ...)
}
\arguments{
\item{gene.presence}{Gene presence/absence matrix.}

\item{sig.genes}{Character vector of the significant genes.}

\item{tree}{A tree object.}

\item{plotted.tree}{A ggtree plot of \code{tree}.}

\item{phylum}{Name of the phylum represented by \code{tree}}

\item{verbose}{Whether to report debugging information (boolean).}
}
\description{
This function wraps single.cluster.plot, running it in a separate process.
This is because there can be problems with memory leaks. For relevant global
options, see the doumentation for that function.
}
