% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{get.pheno.plotting.scales}
\alias{get.pheno.plotting.scales}
\title{Get phenotype plotting scales.}
\usage{
get.pheno.plotting.scales(phenotype, trees, phenoP = NULL, ...)
}
\arguments{
\item{phenotype}{A named vector giving the phenotype for each taxon ID.}

\item{trees}{A list of tree objects.}

\item{phenoP}{An optional value giving the prior probability for the
environment of interest.}
}
\value{
A list of overall limits (\code{limits}), phylum-specific limits
    (\code{phy.limits}), a color scale (\code{colors}), and the zero point
    (\code{zero}).
}
\description{
Some particularly relevant global options are:
\describe{
  \item{which_phenotype}{String. Which phenotype to calculate ("prevalence"
  or "specificity").}
  \item{prev_color_low}{String. When graphing prevalence on a tree, this
  color is the lowest value.} \item{prev_color_high}{String. When graphing
  prevalence on a tree, this color is the highest value.}
  \item{spec_color_high}{String. When graphing specificity on a tree, this
  color is the lowest value (most anti-specific).}
  \item{spec_color_med}{String. When graphing specificity on a tree, this
  color denotes the prior (no association).} \item{spec_color_high}{String.
  When graphing specificity on a tree, this color is the highest value (most
  specific).}
}
}
