% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{harmonize.abd.meta}
\alias{harmonize.abd.meta}
\title{Check metadata and abundance matrix against one another}
\usage{
harmonize.abd.meta(abd.meta, ...)
}
\arguments{
\item{abd.meta}{A list with components \code{mtx} and \code{metadata},
corresponding to a sparse binary presence/absence matrix (see Matrix
package) and a metadata data frame.}
}
\value{
A list of the same form as \code{abd.meta}.
}
\description{
\code{harmonize.abd.meta} compares the abundance matrix to the metadata
matrix to make sure that enough samples are in common between the two to
perform an \emph{phylogenize} analysis, after dropping any singleton datasets
or environments (effects for these cannot be estimated).
}
\details{
Some particularly relevant global options are:
\describe{
  \item{env_column}{Name of metadata column containing environment
  annotations.}
  \item{dset_column}{Name of metadata column containing dataset annotations.}
}
}
