% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{make.results.matrix}
\alias{make.results.matrix}
\title{Convert results into a long (vs. wide) format.}
\usage{
make.results.matrix(results)
}
\arguments{
\item{results}{Output of result.wrapper.plm.}
}
\value{
A single data frame with entries from \code{results}.
}
\description{
Convert results into a long (vs. wide) format.
}
