% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{make.simulated.denoised.data}
\alias{make.simulated.denoised.data}
\title{Wrapper function to simulate data "denoised" by an algorithm like DADA2 or
Deblur.}
\usage{
make.simulated.denoised.data(mtx, tag.length = 100, ...)
}
\arguments{
\item{mtx}{An input abundance matrix.}

\item{tag.length}{Positive integer: how long should amplicon sequence
variants (ASVs) be?}
}
\value{
A list:
  \item{mtx}{The input matrix, with renamed rows.}
  \item{map}{String vector of ASVs, named with the original row names of
    \code{mtx}.}
  \item{n}{Names of the MIDAS taxa to which the rows of \code{mtx} "should"
    map.}
}
\description{
Extra arguments get passed to \code{random.species.from.file}.
}
\keyword{internal}
