% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{NameFaker}
\alias{NameFaker}
\alias{get.taxon.names}
\alias{get.sample.names}
\title{Generate fake names for taxa or samples.}
\usage{
get.taxon.names(n.taxa)

get.sample.names(n.samples)
}
\arguments{
\item{n.taxa}{Number of taxa.}

\item{n.samples}{Number of samples.}
}
\value{
A vector of fake names.
}
\description{
Generate fake names for taxa or samples.
}
\keyword{internal}
