% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{nonequiv.pos.sig}
\alias{nonequiv.pos.sig}
\title{Get vectors of significant genes with positive effect sizes.}
\usage{
nonequiv.pos.sig(results, method = qvals, qcut_sig = 0.05,
  qcut_eq = 0.05, min_fx = 0.25, exclude = NULL)
}
\arguments{
\item{results}{List of result matrices (4 x N).}

\item{method}{Method for performing multiple test adjustment.}

\item{qcut_sig}{Desired q-value cutoff for significance test.}

\item{qcut_eq}{Desired q-value cutoff for equivalence test (N.B.:
significantly equivalent hits will be *excluded*).}

\item{min_fx}{Minimum effect size for equivalence test.}

\item{exclude}{Optional list of character vectors of genes to exclude.}
}
\value{
List of character vectors of hits that were significantly different
    from zero, had positive effect sizes, and not significantly equivalent to
    a minimum effect size.
}
\description{
Get vectors of significant genes with positive effect sizes.
}
