% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{nonparallel.results.generator}
\alias{nonparallel.results.generator}
\title{Fit phylogenetic (or linear) models (single core version, single phylum).}
\usage{
nonparallel.results.generator(gene.matrix, tree, taxa, pheno,
  phylum.name = "TestPhylum", method = phylolm.fx.pv,
  restrict.ff = NULL, remove.low.variance = TRUE,
  use.for.loop = TRUE, ...)
}
\arguments{
\item{gene.matrix}{Gene presence/absence matrix.}

\item{tree}{Phylogeny relating taxa.}

\item{pheno}{Named numeric vector giving phenotype values per taxon.}

\item{phylum.name}{Name of phylum being considered.}

\item{method}{A function that returns a length-2 numeric vector of
effect-size and p-value (see, e.g., \code{phylolm.fx.pv} or
\code{lm.fx.pv}).}

\item{restrict.ff}{Optionally, a character vector giving a subset of
genes to test.}

\item{remove.low.variance}{Boolean giving whether to drop genes that are
always present or always absent in a particular phylum.}

\item{use.for.loop}{Boolean giving whether to use a for loop instead of a
pbapply.}
}
\value{
Named list of p-value and effect-size matrices, one per phylum.
}
\description{
Fit phylogenetic (or linear) models (single core version, single phylum).
}
