% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{pick.env.dset.fx}
\alias{pick.env.dset.fx}
\title{Function for picking taxa to simulate.}
\usage{
pick.env.dset.fx(n.taxa, env.n.affected, dset.n.affected, divided,
  env.mean = 0, env.sd = 0, dset.mean = 0, dset.sd = 0)
}
\arguments{
\item{n.taxa}{Number of taxa to simulate}

\item{env.n.affected}{Number of taxa to be affected by environment effects.}

\item{dset.n.affected}{Number of taxa to be affected by dataset effects.}

\item{divided}{Output of \code{divide.samples}}
}
\value{
A list:
  \item{mtx}{An effect size matrix.}
  \item{fx}{The chosen, randomly generated effect sizes (see
    \code{pick.env.dset.fx}.}
}
\description{
Function for picking taxa to simulate.
}
\keyword{internal}
