% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{plot.labeled.phenotype.trees}
\alias{plot.labeled.phenotype.trees}
\title{Edit a list of plotted trees to add fancy highlight labels.}
\usage{
\method{plot}{labeled.phenotype.trees}(plotted.pheno.trees, phenotype,
  label = "prevalence", stroke.scale = 0.3, units = "\%")
}
\arguments{
\item{plotted.pheno.trees}{A named list of ggtree plots (per phylum).}

\item{phenotype}{Phenotype to plot/label.}

\item{label}{Label to give to the phenotype.}

\item{stroke.scale}{How thick to make the highlight.}

\item{units}{A string appended to each label, used to give units of phenotype.}
}
\description{
This function adds fancy SVG highlight labels to ggtree objects and then
plots them. If there's an error, it will fall back to a regular plot.
}
\details{
Some particularly relevant global options are:
\describe{
  \item{which_phenotype}{String. Which phenotype to calculate ("prevalence"
  or "specificity").}
}
}
