% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{plot.pheno.distributions}
\alias{plot.pheno.distributions}
\title{Plot distributions of a phenotype across phyla.}
\usage{
\method{plot}{pheno.distributions}(phenotype, pz.db, ...)
}
\arguments{
\item{phenotype}{A named vector with the phenotype values for each taxon.}

\item{pz.db}{A database containing a \code{taxonomy} and \code{trees}.}
}
\value{
A ggplot object with the phenotype distribution plotted per phylum.
}
\description{
Some particularly relevant global options are:
\describe{
  \item{which_phenotype}{String. Which phenotype to calculate ("prevalence"
  or "specificity").}
}
}
