% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{prev.addw}
\alias{prev.addw}
\title{Master function to calculate taxon prevalences with additive smoothing.}
\usage{
prev.addw(abd.meta, ...)
}
\arguments{
\item{abd.meta}{A list giving an abundance matrix and metadata.}
}
\value{
An additively-smoothed estimate of taxon prevalences.
}
\description{
Some particularly relevant global options are:
\describe{
  \item{env_column}{String. Name of column in metadata file containing the
  environment annotations.}
  \item{dset_column}{String. Name of column in metadata file containing the
  dataset annotations.}
  \item{which_envir}{String. Environment in which to calculate prevalence or
  specificity. Must match annotations in metadata.}
}
}
