% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{process.16s}
\alias{process.16s}
\title{Map denoised 16S sequence variants to MIDAS IDs using BURST.}
\usage{
process.16s(abd.meta, ...)
}
\arguments{
\item{mtx}{A presence/absence or abundance matrix, with row names equal to
amplicon sequence variant DNA sequences.}
}
\value{
none
}
\description{
\code{process.16s} is a wrapper for other functions that: output sequence
variants to a file; map them using BURST against a reference database of 16S
sequences; then return a list of abundance and metadata values where the rows
of the abundance matrix are now MIDAS IDs.
}
\details{
Some particularly relevant global options are:
\describe{
  \item{in_dir}{String. Path to input directory (i.e., where to look for
  input files).}
  \item{burst_infile}{String. File name of the sequences written to disk and
  then read into BURST.}
}
}
