% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{pz.message}
\alias{pz.message}
\title{Report a message and optionally log it in errmsg.txt.}
\usage{
pz.message(msgtext, ...)
}
\arguments{
\item{errtext}{String: message text.}
}
\description{
Some particularly relevant global options are:
\describe{
  \item{error_to_file}{Boolean. Should pz.error, pz.warning, and pz.message
  output to an error message file?}
}
}
