% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pz.options}
\alias{pz.options}
\title{Set and get options for phylogenize.}
\usage{
pz.options(...)
}
\arguments{
\item{...}{Names of options (to retrieve) or \code{[key]=[value]} pairs (to set).}
}
\description{
Function to set and get global options for the \emph{phylogenize} package.
}
\details{
These options are global because they affect how most of the functions in
\emph{phylogenize} work. Descriptions of these options follow.
}
\section{File input/output and paths}{

\describe{
  \item{out_dir}{String. Path to output directory. Default: "output"}
  \item{in_dir}{String. Path to input directory (i.e., where to look for input files). Default: "."}
  \item{data_dir}{String. Path to directory containing the data files required to perform a \emph{phylogenize} analysis. Default: "./data", but on package load, this default is set to the result of \code{system.file("extdata", package="phylogenize")}.}
  \item{working_dir}{String. Path to directory where relative paths should originate from. Default: \code{"."}}
  \item{abundance_file}{String. Name of abundance tabular file. Default: "test-abundance.tab"}
  \item{metadata_file}{String. Name of metadata tabular file. Default: "test-metadata.tab"}
  \item{biom_file}{String. Name of BIOM abundance-and-metadata file. Default: "test.biom"}
  \item{separate_metadata}{Boolean. For BIOM data, is there a separate tabular abundance table? Default: FALSE}
  \item{input_format}{String. Whether to look for tabular or BIOM-formatted data ("tabular" or "biom"). Default: "tabular"}
  \item{phenotype_file}{String. Name of input file for optional pre-calculated phenotype. Default: ""}
  \item{prior_file}{String. File name of optional pre-computed prior. Default: ""}
  \item{error_to_file}{Boolean. Should pz.error, pz.warning, and pz.message output to an error message file? Default: FALSE}
  \item{biom_dir}{String. Path to BIOM executables. Only used during testing. Default: "/usr/local/bin/"}
  \item{burst_dir}{String. Path where the binary of BURST is found. Default: "/usr/local/bin/"}
  \item{burst_bin}{String. File name of the binary of BURST. Default: "burst12"}
  \item{burst_16sfile}{String. Path to the 16S FASTA database that maps back to MIDAS species. Default: "16s_renamed.frn"}
  \item{burst_infile}{String. File name of the sequences written to disk and then read into BURST. Default: "input_seqs.txt"}
  \item{burst_outfile}{String. File name where BURST writes output which is then read back into \emph{phylogenize}. Default: "output_assignments.txt"}
}
}

\section{Computing phenotypes}{

\describe{
  \item{ncl}{Integer. Number of cores to use for parallel computation. Default: 1}
  \item{type}{String. Type of data to use, either "midas" (shotgun) or "16S" (amplicon). Default: "midas"}
  \item{env_column}{String. Name of column in metadata file containing the environment annotations. Default: "env"}
  \item{dset_column}{String. Name of column in metadata file containing the dataset annotations. Default: "dataset"}
  \item{sample_column}{String. Name of column in metadata file containing the sample IDs. Default: "sample_id"}
  \item{single_dset}{Boolean. If true, will assume that all samples come from a single dataset called \code{dset1} no matter what, if anything, is in \code{dset_column}. Default: FALSE}
  \item{db_version}{String. Which version of the MIDAS database to use ("midas_v1.2" or "midas_v1.0"). Default: "midas_v1.2"}
  \item{which_phenotype}{String. Which phenotype to calculate ("prevalence" or "specificity"). Default: "prevalence"}
  \item{which_envir}{String. Environment in which to calculate prevalence or specificity. Must match annotations in metadata. Default: "Stool"}
  \item{prior_type}{String. What type of prior to use ("uninformative" or "file"). Default: "uninformative"}
  \item{minimum}{Integer. A particular gene must be observed, and also absent, at least this many times to be reported as a significant positive association with the phenotype. Default: 3}
  \item{assume_below_LOD}{Boolean. If TRUE, MIDAS species that are not present are assumed to have a prevalence of zero; if FALSE, they are dropped from the analysis. Default: TRUE}
  \item{linearize}{Boolean. If TRUE, use a regular linear model instead of a phylogenetic linear model. Mostly useful for testing report generation, since the linear model is much faster but returns many more false positives. Default: FALSE}
  \item{burst_cutoff}{Float. Value between 0.95 and 1.00 giving the percent ID cutoff to use when assigning denoised sequence variants to MIDAS species using BURST. Default: 0.985}
  \item{meas_err}{Boolean. Separately estimate measurement error from phenotype variation in the phylogenetic linear model. Default: TRUE}
  \item{min_fx}{Positive double. Effects that are significantly equivalent to this effect size will be excluded from significant positive hits. If zero, the equivalence test will be skipped. Default: 0}
}
}

\section{Graphing}{

\describe{
  \item{treemin}{Integer. A phylum must have at least this many representatives in order to be graphed in the report. Default: 5}
  \item{pctmin}{Integer. A phylum must have at least this percent of observed representatives in order to be graphed in the report. Default: 0.01}
  \item{skip_graphs}{Boolean. If TRUE, skip making graphs in the report, which can be time- and memory-consuming. Default: FALSE}
  \item{prev_color_low}{String. When graphing prevalence on a tree, this color is the lowest value. Default: "black"}
  \item{prev_color_high}{String. When graphing prevalence on a tree, this color is the highest value. Default: "orange2"}
  \item{spec_color_high}{String. When graphing specificity on a tree, this color is the lowest value (most anti-specific). Default: "slateblue"}
  \item{spec_color_med}{String. When graphing specificity on a tree, this color denotes the prior (no association). Default: "gray50"}
  \item{spec_color_high}{String. When graphing specificity on a tree, this color is the highest value (most specific). Default: "tomato"}
  \item{gene_color_absent}{String. When graphing gene presence/absence, this color indicates absence. Default: "black"}
  \item{gene_color_present}{String. When graphing gene presence/absence, this color indicates presence. Default: "black"}
}
}

\section{Memory management}{

\describe{
  \item{pryr}{Boolean. If TRUE, report memory usage when generating the report. Default: FALSE}
  \item{separate_process}{Boolean. When displaying clustered top gene associations alongside a tree colored by phenotype, this flag indicates whether to use a separate subprocess. This allows memory used by clustering to be released back to the operating system immediately. Default: TRUE}
}
}

