% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-enrich.R
\name{qvals}
\alias{qvals}
\title{Wrapper around \code{qvalue} that extracts only q-values. If there is an
error in estimating q-values, will automatically fall back to a
Benjamini-Hochberg-style correction (by setting lambda to zero), finally
returning a vector of NAs if this still does not work.}
\usage{
qvals(x, ...)
}
\arguments{
\item{x}{A vector of p-values.}
}
\value{
A vector of q-values.
}
\description{
Wrapper around \code{qvalue} that extracts only q-values. If there is an
error in estimating q-values, will automatically fall back to a
Benjamini-Hochberg-style correction (by setting lambda to zero), finally
returning a vector of NAs if this still does not work.
}
\keyword{internal}
