% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{read.abd.metadata.biom}
\alias{read.abd.metadata.biom}
\title{Read abundance matrix and metadata (BIOM)}
\usage{
read.abd.metadata.biom(...)
}
\value{
A list with components \code{mtx} (matrix of abundances) and
    \code{metadata} (data frame of metadata).
}
\description{
Read in taxon-by-sample matrix of abundances and metadata (sample
annotations) from a single BIOM-formatted file.
}
\details{
This function uses package-wide options (see \code{?pz.options}),
    which can be overridden using the \code{...} argument. Some particularly
    relevant options are:

\describe{
  \item{in_dir}{String. Path to input directory (i.e., where to look for input files). Default: "."}
  \item{biom_file}{String. Name of BIOM abundance-and-metadata file. Default: "test.biom"}
}
}
\keyword{internal}
