% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-functions.R
\name{read.abd.metadata}
\alias{read.abd.metadata}
\title{Read in abundance and metadata file(s).}
\usage{
read.abd.metadata(...)
}
\value{
A list with components \code{mtx} and \code{metadata}, corresponding
    to a sparse binary presence/absence matrix (see \code{Matrix} package)
    and a metadata data frame.
}
\description{
Read in abundance and metadata, either as one BIOM-format file or as two
tab-delimited files.
}
\details{
This function uses package-wide options (see \code{?pz.options}),
    which can be overridden using the \code{...} argument. Some particularly
    relevant options are:

\describe{
  \item{env_column}{String. Name of column in metadata file containing the
    environment annotations.}
  \item{dset_column}{String. Name of column in metadata file containing the
    dataset annotations.}
  \item{input_format}{String. Whether to look for tabular or BIOM-formatted
    data ("tabular" or "biom").}
  \item{type}{String. Type of data to use, either "midas" (shotgun) or "16S"
    (amplicon).}
}
}
