% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{result.wrapper.plm}
\alias{result.wrapper.plm}
\title{Fit phylogenetic (or linear) models.}
\usage{
\method{result}{wrapper.plm}(phyla, pheno, tree, proteins, clusters,
  method = phylolm.fx.pv, restrict.figfams = NULL,
  drop.zero.var = FALSE, only.return.names = FALSE, ...)
}
\arguments{
\item{phyla}{Character vector giving the names of the phyla.}

\item{pheno}{Named numeric vector giving phenotype values per taxon.}

\item{tree}{Either a single tree covering all taxa, or a list of per-phylum
trees.}

\item{proteins}{Named list of gene presence/absence matrices, per phylum.}

\item{clusters}{Named list of character vectors of taxon IDs, per phylum.}

\item{method}{A function that returns a length-2 numeric vector of
effect-size and p-value (see, e.g., \code{phylolm.fx.pv} or
\code{lm.fx.pv}).}

\item{restrict.figfams}{Optionally, a character vector giving a subset of
genes to test.}

\item{drop.zero.var}{Boolean giving whether to drop genes that are always
present or always absent in a particular phylum.}

\item{only.return.names}{Boolean giving whether to just return the names of
genes to be tested (for debugging).}
}
\value{
Named list of p-value and effect-size matrices, one per phylum.
}
\description{
Fit phylogenetic (or linear) models.
}
