% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-16S.R
\name{run.vsearch}
\alias{run.vsearch}
\title{Run vsearch analysis on a FASTA file of sequences.}
\usage{
run.vsearch(...)
}
\value{
Returns TRUE unless an error is thrown.
}
\description{
\code{run.vsearch} runs, by default, the optimal sequence aligner vsearch
(doi.org/10.5281/zenodo.806850) on a FASTA file, typically one generated by
\code{prepare.vsearch.input}. By changing the \code{vsearch_dir} option, the
aligner vsearch can also be used. If the provided binary name is
not"vsearch", \emph{phylogenize} will assume an old version of vsearch is
being called that doesn't have the ability to search both strands.
}
\details{
Some particularly relevant global options are:
\describe{
  \item{vsearch_infile}{String. File name of the sequences to be read into
vsearch.}
  \item{vsearch_outfile}{String. File name where vsearch writes output
  which is then read back into \emph{phylogenize}.}
  \item{vsearch_dir}{String. Path where the binary of vsearch is found.}
  \item{vsearch_cutoff}{Float. Between 0.95 and 1.00; percent identity
  minimum for alignment results.}
  \item{vsearch_16sfile}{String. Path to the 16S FASTA database that maps
  back to MIDAS species.}
  \item{data_dir}{String. Path to directory containing the data files
required to perform a \emph{phylogenize} analysis. }
}
}
