% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{simulate.counts}
\alias{simulate.counts}
\title{Simulate microbe counts given multinomial probabilities and number of reads.}
\usage{
\method{simulate}{counts}(com, nr)
}
\arguments{
\item{com}{Multinomial probability distribution for all taxa.}

\item{nr}{Read depth.}
}
\value{
Count matrix.
}
\description{
Simulate microbe counts given multinomial probabilities and number of reads.
}
\keyword{internal}
