% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-functions.R
\name{threshold.pos.sigs}
\alias{threshold.pos.sigs}
\title{Filter out genes that are almost always present or absent.}
\usage{
threshold.pos.sigs(pz.db, phy.with.sigs, pos.sig, ...)
}
\arguments{
\item{pz.db}{A database for use with *phylogenize* analyses.}

\item{phy.with.sigs}{A vector of strings giving which phyla had significant
results.}
}
\value{
A single data frame with entries from \code{results}.
}
\description{
Some particularly relevant global options are:
\describe{
  \item{minimum}{Integer. A particular gene must be observed, and also
  absent, at least this many times to be reported as a significant positive
  association with the phenotype.}
}
}
